Move ScreenToWorld calculations to camera
This is coming together as the right place for this to live. I think
having screen size as a member of camera, and potentially synching
changes to the screen/window dimensions back to it would be a decent way
of keeping that stuff in one place.

+++ main.go
@@ -167,7 +167,7 @@ func setup(w, h int) (*system, error) {
 	addTrees(mgr, board)
 
 	hud := game.NewHUD(w, h)
-	camera := NewCamera(float64(w), float64(h))
+	camera := NewCamera(w, h)
 	s := system{
 		render: game.NewRenderer(),
 		combat: NewCombat(mgr, camera),
